refactored database setup code
