+++ WebApiBackend/WebApiBackendTests/UserControllerTest.cs
@@ -1,39 +1,27 @@
-﻿using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 using NUnit.Framework;
 using System;
-using WebApiBackend;
 using WebApiBackend.Controllers;
 using WebApiBackend.Dto;
 using WebApiBackend.Helpers;
 using WebApiBackend.Model;
+using WebApiBackendTests.Helper;
 
 namespace WebApiBackendTests
 {
     class UserTest
     {
-        ServiceDependencyResolver _serviceProvider;
+        private DatabaseSetUpHelper _dbSetUpHelper;
+        private ServiceDependencyResolver _serviceProvider;
         private FlatManagementContext _context;
 
         [SetUp]
         public void Setup()
         {
-            // Builds webhost and gets service providers from web host
-            var webHost = WebHost.CreateDefaultBuilder()
-                .UseStartup<Startup>()
-                .Build();
-            _serviceProvider = new ServiceDependencyResolver(webHost);
-
-            // Resets database to inital state so all tests are isolated and repeatable
-            _context = new FlatManagementContext();
-            _context.Database.EnsureDeleted();
-            _context.Database.EnsureCreated();
-
-            var testDataGenerator = new DevelopmentDatabaseSetup(_context);
-            testDataGenerator.SetupDevelopmentDataSet();
+            _dbSetUpHelper = new DatabaseSetUpHelper();
+            _serviceProvider = _dbSetUpHelper.GetServiceDependencyResolver();
+            _context = _dbSetUpHelper.GetContext();
         }
 
         /// <summary>
